backends/python/api/requirements.txt
Django
psycopg2-binary
djangorestframework
django-cors-headers
PyJWT
gunicorn
environs
